pyproject.toml
[project]
name = "python-backend-filatova"
version = "0.1.0"
description = ""
authors = [
    {name = "DisaniaS",email = "serzh.30303@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic[dotenv,email] (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "python-jose",
    "python-multipart (>=0.0.20,<0.0.21)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "scipy (>=1.15.2,<2.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
